Fix some bugs due to misuse of methods.
Several of the failed test cases that I had observed were actually
due to incorrect arguments being passed into the measure() method
and some of the other methods.

For rules considering a suffix, I was passing in the word / stem
with the suffix under consideration still attached. It seems that
really it's supposed to be a string without the suffix being passed
into the methods. Most of these cases were changed to substrings that
didn't include the suffix, which appears to have fixed a number of the
failed test cases.

Specifically, "rational", "goodness", "adoption", "probate", and "rate"
all appear to get stemmed correctly now.

Unfortunately, a logic error may have been observed regarding how the
rules are implemented with if and elsif structures.

Also, started adding some test cases for some of the methods.