Fix logic error in Step 1b.
There was a logic error for the first rule in Step 1b. Specifically,
the word "feed" was failing. The condition asked if measure("f") > 0 and
"eed" was a suffix of the word ("feed", in this case). measure("f") is
not greater than 0, it wasn't going into that if case. The following
elsif case matched for "feed", and so "feed" was getting stemmed to
"fe". The fix was to change the if to check for the match of "eed" and
then have another if inside to check whether measure("f") > 0.