Fix logic error and speed up performance.
There was a logic error in step_4() that was causing "vitement" to
incorrectly get stemmed as "vitem". Basically, it matched a couple of
the cases and was falling into the wrong one. The fix was to change the
cases from being of the of form measure() > 0 AND match regex to just
matching the regex and then inside having another if to test the
measure() condition. I had thought of doing this anyway to speed up
performance, reasoning that taking a substring in almost every check for
almost every word might be kinda expensive. It appears this was correct
as the stemmer seems to operate much more quickly now. The success rate
has increased to:

(1 - 309 / 30428) = 0.989844879716051