Improve handling of "y" in measure().
The handling of "y" was not quite in alignment with the definitions
described for the algorithm. In my stemmer, "y" was being treated as
both a vowel and a consonant. In reality, "y" should only be being
treated as a vowel when it's preceded by a consonant, as described in
the algorithm. The improper handling of this was causing a few failures
with the test cases.

What I changed is that a copy is made of the word, and a search is done
to see if it contains a consonant followed by a "y". If it does, the
consonant is replaced with "c" and the "y" is replaced with the yen
symbol, "¥".

Then, the definition of a vowel in the regex is updated to include the
usual vowels and "¥", but not "y"

contains_vowel?() was also updated with this handling.

Fixing this has brought the success rate to:

(1 - 0 / 30428) = 1.0